backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pyjwt[crypto]==2.9.0
